backend/pyproject.toml
[project]
name = "gutenberg"
version = "0.1.0"
description = "Office printer gateway: print documents via web GUI or driverless IPP"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "celery==5.5.3",
    "django==5.2.4",
    "djangorestframework==3.16.0",
    "mozilla-django-oidc==4.0.1",
    "pypdf>=6.1.1",
    "python-magic==0.4.27",
    "redis==6.2.0",
]
